new Case10 #1746
try to login after delete wallet
