+++ cypress.json
@@ -1,28 +1,29 @@
 {
-    "chromeWebSecurity": false,
-    "userAgent": "Mozilla/5.0 (X11; Linux x86_64; Cypress) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/78.0.3904.108 Safari/537.36",
-    "defaultCommandTimeout": 60000,
-    "env": {
-        "baseUrl": "https://gooddev.netlify.app/",
-        "mainAccountEmail": "testaccountgd@yahoo.com",
-        "mainAccountUsername": "Automation Test",
-        "mainAccountPhone": "+380988846112",
-        "mainAccountMnemonics": 
-            "spoon blanket example define water coach doll social core cradle fame alone",
-        "wordsForUnsuccessfullLogin": {
-            "withChangedWord": "hello craft urban short color rule time art recipe monster black essay",
-            "withNumbers": "1plunge craft urban short color rule time art recipe monster black essay",
-            "withSymbols": "@plunge craft urban short color rule time art recipe monster black essay",
-            "withChangedLetter": "flunge craft urban short color rule time art recipe monster black essay",
-            "withChangedOrder": "craft plunge urban short color rule time art recipe monster black essay",
-            "withCapitalize": "Craft plunge urban short color rule time art recipe monster black essay"
-        },
-        "additionalAccountEmail": "testaccountgd@yahoo.com",
-        "additionalAccountUsername": "Automation Test",
-        "additionalAccountPhone": "+380988846112",
-        "additionalAccountMnemonics": "spoon blanket example define water coach doll social core cradle fame alone",
-        "emailForCheckingRegistration": "main.test.acc.gooddollar@gmail.com",
-        "usernameForRegistration": "Name Surname",
-        "numberForCheckingRegistration": "+380673001751"
+        "chromeWebSecurity": false,
+        "userAgent": "Mozilla/5.0 (X11; Linux x86_64; Cypress) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/78.0.3904.108 Safari/537.36",
+        "defaultCommandTimeout": 60000,
+        "env": {
+            "baseUrl": "https://gooddev.netlify.app/",
+            "mainAccountEmail": "testaccountgd@yahoo.com",
+            "mainAccountUsername": "Automation Test",
+            "mainAccountPhone": "+380988846112",
+            "mainAccountMnemonics": 
+                "helmet join rate drop will ramp affair fox chimney muscle border leopard",
+            "wordsForUnsuccessfullLogin": {
+                "withChangedWord": "hello craft urban short color rule time art recipe monster black essay",
+                "withNumbers": "1plunge craft urban short color rule time art recipe monster black essay",
+                "withSymbols": "@plunge craft urban short color rule time art recipe monster black essay",
+                "withChangedLetter": "flunge craft urban short color rule time art recipe monster black essay",
+                "withChangedOrder": "craft plunge urban short color rule time art recipe monster black essay",
+                "withCapitalize": "Craft plunge urban short color rule time art recipe monster black essay"
+            },
+            "additionalAccountEmail": "testaccountgd@yahoo.com",
+            "additionalAccountUsername": "Automation Test",
+            "additionalAccountPhone": "+380988846112",
+            "additionalAccountMnemonics": "helmet join rate drop will ramp affair fox chimney muscle border leopard",
+            "emailForCheckingRegistration": "main.test.acc.gooddollar@gmail.com",
+            "usernameForRegistration": "Name Surname",
+            "numberForCheckingRegistration": "+380673001757"
+        }
     }
-}
+
