+++ cypress/PageObjects/EditProfilePage.js
@@ -25,7 +25,7 @@ class EditProfilePage {
   }
 
   get saveButtonText() {
-    return cy.contains('SAVE')
+    return cy.contains(/SAVE/i)
   }
 
   get pageHeader() {
@@ -45,9 +45,8 @@ class EditProfilePage {
   }
 
   get wrongEmailErrorDiv() {
-    return cy.xpath('//*[@id="root"]/div[1]/div/div/div/div[2]/div/div/div/div[2]/div/div/div[1]/div/div[2]', {
-      timeout: 10000,
-    })
+    return contains('Enter a valid format: yourname@example.com')
+    //return cy.xpath('//*[@id="root"]/div[1]/div/div/div/div[2]/div/div/div/div[2]/div/div/div[1]/div/div[2]', { timeout: 10000 })
   }
 
   waitForEditProfilePageDisplayed() {
@@ -70,7 +69,6 @@ class EditProfilePage {
 
   fillUserName(userName) {
     const nameInput = cy.get('input[placeholder="Choose a Username"]', { timeout: 10000 })
-    //.should('contains.value', 'anme')
     nameInput.focus().clear(), { timeout: 1000 }
     nameInput.clear().type(userName), { delay: 400 }
     nameInput.focus().blur()
