+++ cypress/PageObjects/ReceiveMoneyPage.js
@@ -29,7 +29,6 @@ class ReceiveMoneyPage {
   }
 
   get shareLinkButton() {
-    //return cy.get('[data-gdtype]', { timeout: 10000 })
     return cy.contains(/copy link to clipboard/i)
   }
 
@@ -38,7 +37,6 @@ class ReceiveMoneyPage {
   }
 
   get doneButton() {
-    //return cy.get('[data-gdtype="copybutton-done"]', { timeout: 10000 })
     return cy.contains('Done')
   }
