+++ cypress/fixtures/userMnemonicSave.txt
@@ -1,1 +1,1 @@
-solution chunk yard wish tiger kiwi wine acid grocery still mandate balcony
+alley tomorrow stadium able limit naive ill square network bounce century small
