--- cypress/integration/Case1_Sign_up_create.js
@@ -1,43 +1,0 @@
-/* eslint-disable no-undef */
-import StartPage from '../PageObjects/StartPage'
-import SignUpPage from '../PageObjects/SignUpPage'
-import HomePage from '../PageObjects/HomePage'
-
-describe('Test case 1: Create temporary user', () => {
-  it('User to sign up the wallet with correct values', () => {
-    cy.visit(Cypress.env('baseUrl'))
-    localStorage.clear()
-
-    StartPage.createWalletButton.click()
-    SignUpPage.nameInput.should('be.visible')
-    SignUpPage.nameInput.type(Cypress.env('usernameForRegistration'))
-    SignUpPage.nextButton.should('have.attr', 'data-focusable')
-    SignUpPage.nextButton.click()
-    SignUpPage.phoneInput.type(Cypress.env('numberForCheckingRegistration'), { delay: 300 })
-    SignUpPage.nextButton.should('have.attr', 'data-focusable')
-    SignUpPage.nextButton.click()
-    for (let i = 0; i < 6; i++) {
-      SignUpPage.waitForSignUpPageDisplayed()
-      SignUpPage.codeInputs
-        .eq(i)
-        .type(i, { force: true }, { delay: 500 })
-        .should('be.visible')
-    }
-    SignUpPage.emailInput.should('be.visible')
-    SignUpPage.emailInput.type(Cypress.env('emailForCheckingRegistration'))
-    SignUpPage.nextButton.should('have.attr', 'data-focusable')
-    SignUpPage.nextButton.click()
-    SignUpPage.letStartButton.click()
-    SignUpPage.gotItButton.click()
-    HomePage.welcomeFeed.should('be.visible')
-    HomePage.optionsButton.click()
-    HomePage.backupButton.click().should(() => {
-      expect(localStorage.getItem('GD_mnemonic')).to.not.be.null
-      cy.writeFile('../GoodDAPP/cypress/fixtures/userMnemonicSave.txt', localStorage.getItem('GD_mnemonic'))
-    })
-
-    // cy.readFile('../GoodDAPP/cypress/fixtures/userMnemonic.txt').then((text) => {
-    //   localStorage.setItem('GD_mnemonic', text)
-    //  })
-  })
-})
