+++ cypress/plugins/index.js
@@ -17,10 +17,22 @@
 // }
 
 const clipboardy = require('clipboardy')
-module.exports = on => {
+const path = require('path')
+const fakeCameraPath = path.join(__dirname, '/../fixtures/face.mjpeg')
+
+module.exports = (on, config) => {
   on('task', {
     getClipboard() {
       return clipboardy.readSync()
     },
+  })
+
+  on('before:browser:launch', (browser = {}, args) => {
+    if (browser.name === 'chrome') {
+      args.push('--use-fake-ui-for-media-stream')
+      args.push('--use-fake-device-for-media-stream')
+      args.push('--use-file-for-fake-video-capture=' + fakeCameraPath)
+    }
+    return args
   })
 }
